// Assets-Uncapped/front-portal-develop/src/services/api/underwriting/models/BankAccountLabel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-underwriting
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BankAccountLabel
 */
export interface BankAccountLabel {
    /**
     * 
     * @type {string}
     * @memberof BankAccountLabel
     */
    mask?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccountLabel
     */
    label?: BankAccountLabelLabelEnum;
}


/**
 * @export
 */
export const BankAccountLabelLabelEnum = {
    Personal: 'PERSONAL',
    Unrecognised: 'UNRECOGNISED'
} as const;
export type BankAccountLabelLabelEnum = typeof BankAccountLabelLabelEnum[keyof typeof BankAccountLabelLabelEnum];


/**
 * Check if a given object implements the BankAccountLabel interface.
 */
export function instanceOfBankAccountLabel(value: object): value is BankAccountLabel {
    return true;
}

export function BankAccountLabelFromJSON(json: any): BankAccountLabel {
    return BankAccountLabelFromJSONTyped(json, false);
}

export function BankAccountLabelFromJSONTyped(json: any, ignoreDiscriminator: boolean): BankAccountLabel {
    if (json == null) {
        return json;
    }
    return {
        
        'mask': json['mask'] == null ? undefined : json['mask'],
        'label': json['label'] == null ? undefined : json['label'],
    };
}

export function BankAccountLabelToJSON(json: any): BankAccountLabel {
    return BankAccountLabelToJSONTyped(json, false);
}

export function BankAccountLabelToJSONTyped(value?: BankAccountLabel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'mask': value['mask'],
        'label': value['label'],
    };
}